rust: add mortal trait in exchange mod
