+++ rust/src/exchange.rs
@@ -6,7 +6,9 @@
 //
 // Author: Frédéric GARDES <frederic.gardes@orange.com> et al.
 // Software description: This Intelligent Transportation Systems (ITS) [MQTT](https://mqtt.org/) client based on the [JSon](https://www.json.org) [ETSI](https://www.etsi.org/committee/its) specification transcription provides a ready to connect project for the mobility (connected and autonomous vehicles, road side units, vulnerable road users,...).
+
 pub mod message;
+pub mod mortal;
 
 use crate::exchange::message::Message;
 use crate::mobility::position::Position;
